jormungandr: rename script in scenario
