+++ source/jormungandr/jormungandr/scenarios/CMakeLists.txt
@@ -1,0 +1,3 @@
+INSTALL_FILES(/srv/jormungandr/scenarios __init__.py)
+INSTALL_FILES(/srv/jormungandr/scenarios default.py)
+INSTALL_FILES(/srv/jormungandr/scenarios qualifier.py)
