+++ source/jormungandr/jormungandr/scenarios/qualifier.py
