--- source/jormungandr/jormungandr/scripts/CMakeLists.txt
@@ -1,3 +1,0 @@
-INSTALL_FILES(/srv/jormungandr/scripts __init__.py)
-INSTALL_FILES(/srv/jormungandr/scripts default.py)
-INSTALL_FILES(/srv/jormungandr/scripts qualifier.py)
